[#26] Apply code from review
